Update README.md: Ensembles
Updated to reflect the newly added ensemble capability, wherein the output of the LSTM module is the ensemble mean forecast of the output from multiple trained LSTMs. I made relevant changes through the "Trained LSTM Model" section, but the below sections on Dependencies thru Unit Test I did not change (yet), but I think that much of it will be unchanged. I had a few questions though: 

I think we are using AORC forcings, but I do not see those stored in the "Data" folder despite the data section of the readme saying the necessary data lives there. I am not sure where the AORC data lives (in Jonathan's fork of LSTM which is more updated it appears to come from another repo entirely?). Please advise on where I can point a user to the relevant forcings data.